requirements.txt
annotated-types==0.7.0
anyio==4.13.0
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
Deprecated==1.3.1
distro==1.9.0
ecdsa==0.19.2
exceptiongroup==1.3.1
Flask==3.1.3
flask-cors==6.0.2
Flask-Limiter==4.1.1
Flask-SocketIO==5.6.1
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imagekitio==5.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==26.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv==1.2.2
python-engineio==4.13.1
python-jose==3.5.0
python-socketio==5.16.1
requests==2.33.1
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
wrapt==2.1.2
wsproto==1.3.2
